feat(https): use https by default
This PR changes the default behavior of the single-argument
`URLBuilder` constructor. This change ensures that a default
constructed `URLBuilder` has its `useHttps` member variable
set to `true` so that urls are generated for the https scheme.

__This change in behavior represents a breaking change__. Prior to this
change, users could rely on a default-constructed `URLBuilder`'s
`useHttps` member variable to be `false`.

The intention was to test this change. However, with `useHttps` set
to `private` and without reflection, this cannot currently be done.

Here are a few options:
- a new getter could be defined, i.e. `getUseHttps` which feels a little
  clumsy at the moment, __or__
- the interface must be refactored, __or__
- this feature doesn't get tested (unless there's some other way)
  - not-testing is a decision that should be made after deliberation
  - not-testing should only be an option if there are no other
    _reasonable_ options
  - tests before this __breaking change__ should have _passed_
  - tests after this __breaking change__ should have _failed_
  - then a new test should have been written

In order to keep this change appropriately scoped (and modular), any
testing of this feature will be done:

- in a separate PR entirely (possibly after adding a getter or
  refactoring, i.e. private -> public member), _and_
- after discussing potential changes with the SDK Team