chore(release): 2.2.0
This is an unusual commit. It touches more files than most commits
should. This follows up on `feat(https): use https by default #38`
https://github.com/imgix/imgix-java/pull/38, wherein https support
has been enabled by default. After discussing with the SDK Team, we
have agreed to roll up a few things in this commit. We seek the chance
to polish the readme while making the appropriate version changes while
tweaking the build process.

While building the project, we ran into issues getting the same
results machine to machine. This commit aims to solve these issues
now: a new release and fresh build (devoid of deprecation
warnings) means putting our "best foot forward" for `2.2.0`.

If this commit is merged it means that the new build setup works
for everyone on the SDK Team. After this commit is merged there
will be another request made to update the instructions to build
this project from source to help our community.

Summary of Changes:

- updates the gitignore to guard against pushing `mistakes`
- updates the changelog to reflect current version
- updates readme to:
  - reflect current version
  - remove adverbs ending in `ly`; they _essentially_ waste space
  - fix a minor formatting issue in the table of contents
- updates `build.gradle` to achieve reproducible builds __and__
- to follow the guidance of deprecation warnings:
  - no `testCompile` use `testImplementation`
  - no `apply plugin: 'java'` place these in `plugins {...}`
  - no `apply plugin: 'maven'`, just use `maven-publish`
  - `classifier = 'javadoc'` is a deprecated feature--removed
  - `classifier = 'sources'` is a deprecated feature--removed!
- updates gradle build files via:
 - `$ gradle wrapper --gradle-version 6.3`