build: upgrade gradle usage to prep for 7.0
The following applies to Gradle 6.3 and this commit updates the current
Gradle version from **5.4** to **6.3**. NOTE: This commit DOES NOT
UPDATE the following,
```
sourceCompatibility = 1.6
targetCompatibility = 1.6
```
The idea behind the upgrade is to:
1. Stay up to date with the community
2. Comply with deprecation warnings
3. Prepare to automate our java build process (like we've done in Php)

The current build process yields deprecation warnings that will be
errors in Gradle 7.0. In anticipation of this, and in order to update
our build process to be reproduced machine-to-machine, this PR makes
the following changes:

- upgrade with the gradle wrapper to get better IDE support
  - `$ gradle wrapper --gradle-version 6.3 --distribution-type all`

**Deprecation Fixes**:
This style is deprecated...
```text
apply plugin: 'java'
apply plugin: 'maven'
```
In favor of:
```
plugins {
    --snip--
    id 'java'
    // Note: the use of 'maven' alone is deprecated
    // guidance is to use the following
    id 'maven-publish'
}
```
This style has been deprecated:
```
testCompile group: 'junit', name: 'junit', version:'4.12'
```
In favor of this:
```
testImplementation group: 'junit', name: 'junit', version:'4.12
```
- `classifier = 'javadoc'` is a deprecated feature
  - prefer `archiveClassifier.set("javadoc")`
- `classifier = 'sources'` is a deprecated feature
  - prefer `archiveClassifier.set("sources")`

There is more work to be done, but this is a start.

**The Following Pass**:
- [x] `$ gradle build`
- [x] `$ gradle test`
- [x] `$ gradlew build`
- [x] `$ gradlew test`