feat: prevent push mistakes (#41)
This commit updates `.gitignore` to guard against pushing `mistakes/`.

To accommodate more contributors, a few editor specific cases have been
added to `.gitignore`. This way, we can reduce the chances of pushing
things we don't want to and focus on the task at hand. Examples of
these things are build files, `dist/` and `bin/` directories, etc.